Split key on regex when normalizing for lookup
This addresses some unexpected behavior when handling certain keys that match the provided (or default - '.') separator.

If a key is matches the separator, `I18n.t('.')` will dump the `@translations` hash from the backend for the provided locale.

If the key ends with the separator, `I18n.t('foo.bar.baz.')` will not use the trailing `.` when looking up a translation, resulting in a `I18n::MissingTranslationData` error.

Both cases are addressed by splitting on a regex, instead of the separator itself. This incurs a slight performance cost (as well as maintenance overhead as regexes are not particularly human-readable).